Translation changes and additions
I made changes to the translations. Changed capitalization theming. Added en.ft = "ft" as an abbreviation for featured artist. I don't know if it may be useful or not. A few items I am not certain as to the best way to name some of the options. I will need to use the player a bit before knowing better how to name certain options. You used a few amusing choices in the translations, like describing a track to be voted as 'crap'. LOL!

I hope this helps some.